requirements.txt
allure-pytest
Appium-Python-Client==1.2.0
bs4
influxdb
influxdb-client
pandas
paramiko
perfecto-py37
pyshark
pytest==6.2.2
requests
scp
selenium
tip-wlan-cloud
xlsxwriter